Revert "refactor(election): move election widget above news marquee"
This reverts commit 9b5691069021f574cc4d6ffb398d8b1d978b0231.

+++ lib/models/election/candidate.dart
@@ -21,7 +21,7 @@ class Candidate {
       name: json['name'],
       party: json['party'],
       votes: json['tks'],
-      percentageOfVotesObtained: json['tksRate'].toDouble(),
+      percentageOfVotesObtained: json['tksRate'],
       elected: json['candVictor'],
     );
   }
